ThRead/library_back/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.2
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.1
Django==4.2.21
django-allauth==0.61.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
exceptiongroup==1.2.2
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==2.0.2
oauthlib==3.2.2
openai==1.61.1
Pillow==10.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.3.0
Wikipedia-API==0.8.1
